Fix click.pointerId for touch events sent to subframes.
Blink's pointerId generation is attached to low-level PointerEvent
handling (like pointerdown/up/move) which stores the data in the
local-frame-root.  However, to find a touch click's pointerId (which is
needed in an asynchronous manner), we have been accessing at the
subframe's state because of a historic crack in Blink EventHandler code
(https://crbug.com/449649).  This CL fixes the id-data lookup to access
local-frame-root's state instead.

Bug: 1264930
Change-Id: I9d2a76dd18c63086afd253575f99bf7f4f26bc7d