[Fixit] Make sandbox + COOP break the opener.
Sandbox flags and COOP cannot currently live on the same page. This was
decided when first implementing COOP. If COOP and sandboxing flags are
present on the same top level window, we return an error page instead,
and COOP has no effect.

Exploits have been discovered recently about abusing an opened window
and history to be able to guess the cross-origin url using the history
API. See:
https://bugs.chromium.org/p/chromium/issues/detail?id=1208614

Error pages were left out from the fix, because it would be inelegant
to insert error pages into history after a successful reload. That means
error pages can potentially be used to guess a cross-origin URL.

If in general this is acceptable, in our case its worse for two reasons:
- COOP is used to protect pages that are usually more valuable or hold
  important data.
- It is possible to reach an error page after response without relying
  on timing, etc.

Therefore we'd like to sever the opener when we fail after receiving a
response, if the target page did set COOP. Basically we'd like to change
the spec to enforce COOP even when we fail because of COOP+sandbox.

Spec discussion can be found here:
https://github.com/whatwg/html/issues/7345

On the implementation side, this patch adds the WPTs necessary to the
spec change, and the modified behavior.

We add a setter to the swap parameter to be able to only activate that
without breaking the COOP: Unsafe-none for error pages invariant. This
applies to final responses and redirects.

This does not alter the reload behavior, as this only happens for cases
where reloads would not be successful anyway.

Bug: 1256823
Change-Id: Ic79623a2b752608bc46a031d1f567308441d42e0